refactor: rename package to ice-stream and remove mocks
