+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "actions-package"
+name = "ice-stream"
 version = "0.1.0"
 description = "A basic skeleton Python 3.12 package with pytest support"
 readme = "README.md"
