--- src/actions_package/__init__.py
@@ -1,10 +1,0 @@
-"""
-actions-package: A basic skeleton Python 3.12 package with pytest support
-"""
-
-__version__ = "0.1.0"
-
-from .hello import hello_world
-from .azure_storage import AzuriteStorageClient
-
-__all__ = ["hello_world", "AzuriteStorageClient"]
