+++ src/ice_stream/__init__.py
@@ -1,0 +1,4 @@
+"""ice-stream: utilities for streaming Icechunk data."""
+
+__all__ = ["__version__"]
+__version__ = "0.1.0"
